@@ -1,7 +1,6 @@
 #include <iostream>
 #include <memory>
-
-#include <boost/optional.hpp>
+#include <optional>
 
 #include "baldr/streetname.h"
 #include "baldr/streetnames_us.h"
@@ -16,16 +15,17 @@ StreetNamesUs::StreetNamesUs() : StreetNames() {
 
 StreetNamesUs::StreetNamesUs(const std::vector<std::pair<std::string, bool>>& names) {
   for (auto& name : names) {
-    this->emplace_back(std::make_unique<StreetNameUs>(name.first, name.second, boost::none));
+    this->emplace_back(std::make_unique<StreetNameUs>(name.first, name.second, std::nullopt));
   }
 }
 
 StreetNamesUs::StreetNamesUs(const google::protobuf::RepeatedPtrField<valhalla::StreetName>& names) {
   for (auto& name : names) {
-    boost::optional<baldr::Pronunciation> pronunciation =
-        boost::make_optional(name.has_pronunciation(),
-                             baldr::Pronunciation{name.pronunciation().alphabet(),
-                                                  name.pronunciation().value()});
+    std::optional<baldr::Pronunciation> pronunciation =
+        name.has_pronunciation()
+            ? std::make_optional(
+                  baldr::Pronunciation{name.pronunciation().alphabet(), name.pronunciation().value()})
+            : std::nullopt;
 
     this->emplace_back(
         std::make_unique<StreetNameUs>(name.value(), name.is_route_number(), pronunciation));
