@@ -44,7 +44,7 @@ protected:
     test::customize_historical_traffic(use_tracks_map.config, [](baldr::DirectedEdge& e) {
       e.set_free_flow_speed(40);
       e.set_constrained_flow_speed(40);
-      return boost::none;
+      return std::nullopt;
     });
   }
 };
