@@ -46,8 +46,13 @@ private:
 class IViterbiSearch;
 
 // TODO test it
-class StateIdIterator : public std::iterator<std::forward_iterator_tag, StateId> {
+class StateIdIterator {
 public:
+  using iterator_category = std::forward_iterator_tag;
+  using value_type = StateId;
+  using difference_type = std::ptrdiff_t;
+  using pointer = StateId*;
+  using reference = StateId&;
   StateIdIterator(IViterbiSearch& vs,
                   StateId::Time time,
                   const StateId& stateid,
