@@ -1683,6 +1683,24 @@ struct OSMWay {
   }
 
   /**
+   * Sets the lit state
+   *
+   * @param lit whether the way is lit.
+   */
+  void set_lit(const bool lit) {
+    lit_ = lit;
+  }
+
+  /**
+   * Get the lit state.
+   *
+   * @return bool
+   */
+  bool lit() const {
+    return lit_;
+  }
+
+  /**
    * Get the names for the edge info based on the road class.
    * @param  ref              updated refs from relations.
    * @param  name_offset_map  map of unique names and refs from ways.
@@ -1825,7 +1843,8 @@ struct OSMWay {
   uint16_t use_sidepath_ : 1;
   uint16_t bike_forward_ : 1;
   uint16_t bike_backward_ : 1;
-  uint16_t spare2_ : 4;
+  bool lit_ : 1;
+  uint16_t spare2_ : 3;
 
   uint16_t nodecount_;
 
