@@ -1,5 +1,5 @@
 #pragma once
 
 #define VALHALLA_VERSION_MAJOR 3
-#define VALHALLA_VERSION_MINOR 2
+#define VALHALLA_VERSION_MINOR 3
 #define VALHALLA_VERSION_PATCH 0
